pyproject.toml
[tool.poetry]
name = "figma-backend"
version = "0.1.0"
description = ""
authors = ["Hypercodding <mohammadhammad302@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1.3"
supabase = "^2.5.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
